requirements.txt
anthropic>=0.40.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
scipy>=1.11.0
